+++ src/main/java/com/inari/firefly/physics/movement/SimpleStepIntegrator.kt
@@ -5,7 +5,7 @@ import kotlin.math.abs
 
 object SimpleStepIntegrator : Integrator {
 
-    override fun integrate(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Float) {
+    override fun integrate(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Long) {
         val velocity = movement.velocity
 
         if (movement.onGround) {
@@ -13,10 +13,9 @@ object SimpleStepIntegrator : Integrator {
                 velocity.dy = 0f
             return
         }
-       velocity.dy =
-            velocity.dy + abs( (velocity.dy / movement.mass - 1f ) * 0.2f )
+       velocity.dy = velocity.dy + abs( (velocity.dy / movement.mass - 1f ) * 0.2f )
     }
 
-    override fun step(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Float) =
+    override fun step(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Long) =
         transform.move(movement.velocity.dx, movement.velocity.dy)
 }
